AD-HOC refactor (docker-compose): Remove the port binding for MySQL
Previously, in docker-compose the mysql 3306 port was bound to the
localhost adapter at 3306, so we could easily make queries of this
host by connecting to it at 127.0.0.1. However, there are few reasons
to connect directly to MySQL on a regular basis. Where it is necessary,
the developer can find the IP assigned to the MySQL container
by executing:

    docker inspect dockercompose_db_1

And looking for the IP node, then directly connecting to that.
This resolves an issue where in the build environment port 3306 was
already bound.