AD-HOC feat (build): Add CI/CD PGP private key
Access to the various services that this build will influence
needs credentials, such as access to the Magento API and access
to quay.io. git-crypt was previously initialised on this
repository so that secrets can be kept.

This commit allows the secret key to be imported into the gpg
key store, where it can be later used by git-crypt.