AD-HOC refactor (Deployment): Remove secret spec from repo
Principally, there exists a chicken-and-egg problem with the
installation of Magento. In order to deploy it idempotently, we
need to presume that it is installed. However, to deploy it correctly
git commit -F - <<EOF
AD-HOC refactor (Deployment): Remove secret spec from repo

Principally, there exists a chicken-and-egg problem with the
installation of Magento. In order to deploy it idempotently, we
need to presume that it is installed. However, during installation
it generates its own secrets (cryptographic key) which we cannot
know before it installs, and thus cannot deploy idempotently

MySQL solves this by creating its on secret during initialisation.
A similar approach is going to be used in this case. However,
this necessitates the removal of the secret spec.